TST: Clean up pickle compression tests
