+++ pandas/tests/frame/test_to_csv.py
@@ -925,22 +925,23 @@ class TestDataFrameToCSV(TestData):
                         [12.32112, 123123.2, 321321.2]],
                        index=['A', 'B'], columns=['X', 'Y', 'Z'])
 
-        with ensure_clean() as filename:
+        if compression != "zip":
+            with ensure_clean() as filename:
 
-            df.to_csv(filename, compression=compression)
+                df.to_csv(filename, compression=compression)
 
-            # test the round trip - to_csv -> read_csv
-            rs = read_csv(filename, compression=compression, index_col=0)
-            assert_frame_equal(df, rs)
+                # test the round trip - to_csv -> read_csv
+                rs = read_csv(filename, compression=compression, index_col=0)
+                assert_frame_equal(df, rs)
 
-            # explicitly make sure file is compressed
-            with tm.decompress_file(filename, compression) as fh:
-                text = fh.read().decode('utf8')
-                for col in df.columns:
-                    assert col in text
+                # explicitly make sure file is compressed
+                with tm.decompress_file(filename, compression) as fh:
+                    text = fh.read().decode('utf8')
+                    for col in df.columns:
+                        assert col in text
 
-            with tm.decompress_file(filename, compression) as fh:
-                assert_frame_equal(df, read_csv(fh, index_col=0))
+                with tm.decompress_file(filename, compression) as fh:
+                    assert_frame_equal(df, read_csv(fh, index_col=0))
 
     def test_to_csv_compression_value_error(self):
         # GH7615
